bcm2835-bootfiles: Do not deploy debug binaries in the resinOS rootfs
Do not add in resinOS the files fixup_db.dat and start_db.elf which are esentially versions
of fixup_x.dat and start_x.elf with debug assertions enabled.

By removing them we free up about 4.8 MB from resin-boot (having less free space in resin-boot
makes it probable that hostapps update will fail when one of the files in this boot partition
grows in size in the future, such as the kernel for example; when one file in resin-boot is bigger
than the available free space, hostapps update will error out because it tries to do an atomic update
of the bootfiles).

Signed-off-by: Florin Sarbu <florin@resin.io>