linux-raspberrypi_%.bbappend: Blacklist rtl8192cu wifi driver
Currently the raspberry pi kernel builds both the out of tree 8192cu kernel module
and the in-tree rtl8192cu kernel module. We are shipping both drivers in our OS and
this has the effect of the Realtek Adapter 8192cu to not work at all.

So we need to blacklist one of these modules to make this wifi adapter work.
Considering the comments on the raspberrypi official repository,
https://github.com/raspberrypi/linux/issues/1866
we will blacklist the upstream driver and stick with the out of tree module which seems
to yield better performance.

Signed-off-by: Florin Sarbu <florin@resin.io>