linux-raspberrypi: Revert patch which disables memory cgroup controller
In resinOS the supervisor needs the ability to manage memory allocated to the
containers. In order to do this, it needs the cgroup memory controller not only
compiled (available) but as well enabled as raspberrypi kernel has a custom
patch which disables it by default:

https://github.com/raspberrypi/linux/commit/ba742b52e5099b3ed964e78f227dc96460b5cdc0

This change reverts the above patch which disables the memory cgroup
controller.

Signed-off-by: Andrei Gherzan <andrei@resin.io>