Add udev rule to switch U-Blox modem from RNDIS to ECM mode
The board is unable to connect to resin VPN when using an U-blox Toby-L210 modem in RNDIS mode.
This commit changes the modem mode from RNDIS to ECM, when the modem is detected in RNDIS mode.

This udev rule will trigger just once, and it will restart the modem repeatedly until it gets into ECM mode.
When the modem succesfully resets in ECM mode, the board will reboot and the udev rule will not be triggered
anymore because the USB ProductId of the modem changes.
The modem will remain in ECM mode after power cycling.

This commit is a workaround, and it doesn't treat the root cause.
The VPN connection works on ResinOS v2.3.0 even if the modem is in RNDIS mode.

Signed-off-by: Sebastian Panceac <sebastian@resin.io>