u-boot: Add Automated Rollback support to raspberry pi
To enable Automated Rollback support, we have to make u-boot read
upgrade_available from resinOS_uEnv.txt and we have to make u-boot
store the bootcount variable in a file called bootcount.env in the
boot partition.
The patches do the following
- Enable the bootcount config options.
- Add support to the ext bootcount driver to write bootcount
in the fat partition.
- u-boot runs altbootcmd if bootlimit is exceeded.
We handle this manually in env_resin.h where the resin_root_part is
switched. Hence use the same bootcmd for altbootcmd.

Changelog-entry: Add Automated Rollback support
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@resin.io>