u-boot: Change CONFIG_AUTOBOOT_DELAY_STR to 'pause'
Currently, CONFIG_AUTOBOOT_DELAY_STR is set to the space character.
However, in cases where there is another piece of hardware connected
to the Pi's UART, it is likely that this connected hardware sends the
space char, thus putting the autoboot process to a halt. So we change
this to be the 'pause' string instead to reduce likeliness of an
accidental delay happening.

Changelog-entry: Change u-boot's CONFIG_AUTOBOOT_DELAY_STR to 'pause'
Signed-off-by: Florin Sarbu <florin@resin.io>