hostapp-update-hooks: Add hook to clean up kernel in boot partition
The kernel in the root partition in the /boot folder is now used to
boot resinOS. After a HUP, there will be a redundant vmlinuz file
left in the boot partition.

This hooks tests for its presence and removes the redundant file.

Changelog-entry: Add hook to remove kernel from boot parition after a HUP
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@resin.io>