u-boot: Disable autoboot delay option and enable CTRL^C stop feature
We do not have at this point an usecase for delaying the autoboot. In
addition, we have witnessed a case when an IC connected to uart was
delaying the autoboot by sending lots of data and thus putting the
entire autoboot to a halt.

We also change the stop string to be the ASCII code for Vertical Tab
and enable the alternate option to stop autoboot by pressing CTRL^C.
In such a case, the actual viable option to stop autoboot remains
CTRL^C but u-boot also needs CONFIG_AUTOBOOT_STOP_STR defined for
CONFIG_AUTOBOOT_KEYED_CTRLC to function correctly so that is why we
set the stop string to VT.

Changelog-entry: Disable u-boot autoboot delay option and enable CTRL^C autoboot stop feature
Signed-off-by: Florin Sarbu <florin@resin.io>