u-boot: Disable rollbacks
The current integration in uboot uses the ext bootcount driver which
doesn't use upgrade_available. The entire rollback support design is
currently based on this variable (upgrade_available) so if that is not
taken into consideration the board will increment bootcount with every
boot breaking the board after the first boot. Disable for now this
feature.

Changelog-entry: Disable rollbacks
Signed-off-by: Andrei Gherzan <andrei@balena.io>