hostapp-update-hooks: Add config.txt quirk for revpi-core-3
New balenaOS for revpi-core-3 will not boot if the config.txt
has an older dtoverlay=mmc entry.
Therefore we must add a hostapp update hook to remove this entry.

We name this quirk 9999-bootfiles to be sure it is run at the very
end so that no other hooks that may fail would run after it and as
such ending up rebooting in the older OS but with config.txt modified
by this quirk hook.

Changelog-entry: Add config.txt entry removal quirk for revpi-core-3
Signed-off-by: Florin Sarbu <florin@balena.io>