bluez5_%.bbapend: Add BT dependency for the Balena machine
The pi-bluetooth gets chained in through
a dependency in bluetooth, in thud. This dependency
is defined in bluez5 and matches the machine.
Currently, the raspberrypi BSP only does it for
raspberrypi zero wireless and raspberrypi3.
To adapt this for the balena machine setup (which has
raspberrypi and raspberrypi zero wireless support in the
same machine) the same dependency needs to be added.

Changelog-entry: Add bluetooth dependency for the raspberrypi machine
Signed-off-by: Vicentiu Galanopulo <vicentiu@balena.io>