linux-raspberrypi: Repurpose Marvell pwrseq driver to support balenaFin
The original version had a reset pin, which does not exist for this chipset. The
related dts property is purged from the source code.

A `post_power_on` hook is now used instead of `pre_power_on`. Specifications
for all Marvel chips require the PDn pin to be asserted AFTER the chip is powered
on. The original implementation using `pre_power_on` was incorrect.

The balenaFin device tree overlay is adjusted to be using the patched driver.

Changelog-entry: Repurpose Marvell pwrseq driver to support balenaFin
Signed-off-by: Zahari Petkov <zahari@balena.io>