linux-raspberrypi: Disable bcm2708_fb driver
This driver is built-in so we disable it's init.
We need to disable this kernel driver in order to avoid concurrent
access to the GPU from both the kernel bcm2708_fb driver and u-boot's
simplefb driver. (See https://github.com/raspberrypi/linux/issues/3139)

Changelog-entry: Disable bcm2708_fb driver (kernel 4.19.66) so we have working graphics
Signed-off-by: Florin Sarbu <florin@balena.io>