meta-balena-raspberrypi:local.conf.sample: Do not enable vc4-kms-v3d
We leave this overlay enabled only for the 64 bits machines we support
(RaspberryPi 3 64 bits and RaspberryPi 4 64 bits) so that we are
in sync with already deployed devices. We will enable it here again
eventually when this setting is managed by the supervisor in the
saved target state.

Changelog-entry: Do not enable the vc4-kms-v3d overlay for non-64 bits machines devices
Signed-off-by: Florin Sarbu <florin@balena.io>