linux-raspberrypi: Avoid kernel panic if booting with Sensehat
If a framebuffer console is used with the
Sensehat's 8x8 display framebuffer, various
issues will appear when attempting to draw
images on it.

This small display is not meant for console,
therefore fbcon should not takeover the fb
created by its' driver.

This issue does not occurr if another
framebuffer is already present and used,
as console is not switched to the new one.

Changelog-entry: linux-raspberrypi: Avoid kernel panic if booting with Sensehat
Signed-off-by: Alexandru Costache <alexandru@balena.io>