hostapp-update-hooks: Handle modules in old hostOS for pieeprom
When kernel is updated, modules cannot be loaded from
the hook running in a container with the new hostOS,
to which we are updating.

Switch module handling to old hostOS so modules version
match current running kernel.

Also, ensure failures in updating EEPROM do not
trigger failure of HUP.

Changelog-entry: hostapp-update-hooks: Handle modules in old hostOS for pieeprom
Signed-off-by: Alexandru Costache <alexandru@balena.io>