u-boot: query customer OTP register in rpi3
This patch adds support for querying the firmware
for one customer OTP register, as an alternative
for the case where bootmode register cannot be
inspected due to missing TAG implementation in firmware.

Since documentation does not seem to specify that TAG,
asked question in upstream wheter it is available or not.

This patch is therefore incomplete, will suffer changes
and therefore should not be merged as-is.

Changelog-entry: u-boot: query customer OTP register in rpi3
Signed-off-by: Alexandru Costache <alexandru@balena.io>