fincm3.conf: Add flag for sd8887-mrvl from source or from binaries
We will now compile sd8887-mrvl from source only if the
SD8887_FROM_SOURCE variable is set. This allows for other
people not having access to the driver source to fallback to
using the prebuilt binaries we offer.
For build machines having access to the source repos, the source
compilation will be forced by adding "-a SD8887_FROM_SOURCE=1"
to barys.

Signed-off-by: Florin Sarbu <florin@balena.io>