u-boot: Prevent blocking writes to serial console
It appears that when the console is incorrectly configured
in the dtb, booting will get stuck while trying to write
chars to the uart.

The 16bit for the counter seems the most appropriate, as
booting time is insignificantly impacted in the worst case,
while no chars from logs appear to be lost if the uart is
configured correctly.

Changelog-entry: u-boot: Prevent blocking writes to serial console
Signed-off-by: Alexandru Costache <alexandru@balena.io>