sd8887-mrvl-prebuilt.bb: Fix bluetooth and mlan loading order
We previously introduced dependency loading between bt8xxx and mlan
but yocto did not generate the modprobe.d file as expected so now let's
correct that by defining the mlan dependency on bluetooth in a separate
file altogether.

Also fix a packaging issue which made the build with the prebuilts
to not contain the firmware and the modprobe.d conf files.

Signed-off-by: Florin Sarbu <florin@balena.io>