[TEST] linux-raspberrypi: Modifications for kexec support
This is only for test purposes as only the virtual/balena-kboot kernel
needs to have kexec support.

* Allow spin-table cpus to kexec

As the kernel hotplug subsystem does not support cpus that use the spin-
table policy to bring up secondary CPUs, kexec is not supported for these
platforms apart from the kdump use case.

This patch implements a hack to allow spin-table cpus to leave the kernel
allowing kexec to work. This should be safe for non-SMP modes.

* Add a memory node for the raspberrypi4

This is usually dynamically added by bootloaders, but in the case of using
a kexec boot it needs to be encoded in the device tree.

Signed-off-by: Alex Gonzalez <alexg@balena.io>