linux-kunbus: Update to version 4.19.95
We update to the Kunbus recommended kernel version.
We remove the Netflix vulnerability patches which are now merged in
kernel 4.19.95. Also, we remove
0001-Revert-cgroup-Add-kernel-param-to-enable-memory-grou.patch since
it's not needed anymore and we also remove
0001-Add-poe-overlay-for-compatiibility.patch since that overlay is now
in present in the kunbus kernel source tree.
We also removed linux-raspberrypi.inc and we are basing on the
equivalent file from the BSP.

Changelog-entry: Update the Revolution Pi kernel to version 4.19.95
Signed-off-by: Florin Sarbu <florin@balena.io>