u-boot: Fix Pi3 prod image not booting with u-boot 2020.01
U-boot has been updated in the upstream meta-raspberrypi
repository, and this new version freezes at boot if
enable_uart is not set in config.txt, like for instance
in production images. We've seen and fixed this issue already
for the Pi4 which already used a newer uboot, so let's apply
it for all boards.

Signed-off-by: Alexandru Costache <alexandru@balena.io>