meta-balena-raspberrypi/linux: fix ethernet cable detection
Ethernet cable connect/disconnect events are not detected when a
device is booted with the ethernet cable already connected.

This is a known issue:
https://github.com/raspberrypi/linux/issues/2937

This commit applies an upstream patch that is already in the
raspberrypi-linux rpi-5.4.y and rpi-5.10.y branches.

Change-type: patch
Connects-to: #605
Changelog-entry: meta-balena-raspberrypi/linux: fix ethernet cable detection
Signed-off-by: Mark Corbin <mark@balena.io>