bluez5: Use bluez5 v5.58 from meta-balena
Removed are the bluez5 v5.56 definitions together with two patches
that were backported here and are already included of v5.58.

The nebra-hnt specific setting of DeviceID to false is done in
meta-balena now for all device types.

Changelog-entry: Remove obsolete bluez5 patches and settings that are now included in bluez v5.58 from meta-balena
Signed-off-by: Zahari Petkov <zahari@balena.io>