bootfiles: Use rpi firmware revision set in the previous release
One of the testbot DUT devices, a Pi4 Rev 1.1 used on worker
with uuid 06c77b8 does not boot properly with the newer
firmware and it's the only one to exhibit this behavior in
which it reboots continuously about 30s into booting.

This behavior however is not reproducible with local
devices, only with that particular device in the rig.

Signed-off-by: Alexandru Costache <alexandru@balena.io>