u-boot: Ensure uart retry count is stored in data section
The Pi3 32bit currently refuses to boot if
gpu_mem is set to 396MB. This seems to be caused
by a variable we use in u-boot to avoid locking
while writing to the uart, if enable_uart is
disabled in config.txt.

Having the putc_retry variable uninitialized
meant that it was part of the data section,
however this no longer seems to be the case and
we need to allocate it in .data explicitly.

Changelog-entry: u-boot: fix Pi3 booting with large gpu_mem
Signed-off-by: Alexandru Costache <alexandru@balena.io>