Change Fin WiFi/BT pins drive strength to 6mA
Currently pins 34-39 that are used for the wireless chip on balenaFin
boards are not being specified in the Fin dt-blob. This gives them
a default drive strength of 2mA, which causes signal integrity issues.

This commit changes them to 6mA, which is two steps above the current
setting.

Change-type: patch
Signed-off-by: Zahari Petkov <zahari@balena.io>