rpi-bootfiles.bbappend: Make the firmware filtering generic
We do this so we don't need to always increase this list we check each time
we add new raspberrypi4-64 based machines. Instead, we rely on the fact that
all new RPI4 based machines will inherit the raspberrypi4-64 machine.

Changelog-entry: Make the boot firmware filtering for rpi4 based hw generic
Signed-off-by: Florin Sarbu <florin@balena.io>