rpi-eeprom: remove module loading/unloading
Loading modules on demand is problematic in the case of rollbacks, as
the running kernel does not match the rolled back root filesystem
modules.

Replace with building in the SPI support and binding/unbinding the
driver when the specific dynamic configuration is changed.

Changelog-entry: use built-in SPI drivers for EEPROM access
Signed-off-by: Alex Gonzalez <alexg@balena.io>