balena-image: increase boot partition size
When performing a hostOS update we perform atomic writes of individual
files, that is, first we copy a new file, and then we rename.

The rename operation is atomic on some filesystems like ext4, but on
vfat it has not been atomic until Linux kernel v6.0.

Hence, balenaOS uses a `fatrw` library that keeps a checksummed copy of
the file until the rename completes and can be verified. Unfortunately,
for hostOS updates `fatrw` will fallback to a non-atomic rename if there
is not enough space to keep a checksummed copy of the file. This will
stop being a problem once a >6.0 kernel is used.

Now, a secure boot enabled CM4/RPI4 required a single boot.img that
contains all the essential boot firmware. This contains the balena
bootloader and its initramfs that needs packages to mount encrypted
file systems. Basically, the size of the balena booloader increases to
around 50M, and the size of boot.img to around 60M, so we need at least
double that in order to perform atomic updates of this file.

Changelog-entry: increase the boot partition size for CM4 to support secure boot
Signed-off-by: Alex Gonzalez <alexg@balena.io>