initramfs-framework: revpi:remove migrator module
The Revolution Pi are particular devices as they define an
internal storage in the contract, but they do not use a flasher
image and instead rely on the bootloader's exposing the internal
storage.

However, resin-init-flasher has a built time check for all devices
that define internal storage to define INTERNAL_DEVICE_KERNEL, so the
build fails when initramfs-framework builds the resin-init-flasher
dependency.

The lack of a `jq` dependency on earlier versions of meta-balena
meant that the check was non-operative, but it started failing when
the dependency was added.

This commit removes the migrator module from being built and included
in the initramfs.

Changelog-entry: remove migrator module for revpi family
Signed-off-by: Alex Gonzalez <alexg@balena.io>