hostapp-update-hooks: customize for secure boot
Duplicate 1-bootfiles into 2-encbootfiles so that the bootfiles are
installed in two passes, one for the non-encrypted boot partition and
a second one for the encrypted boot partition.

The 1-bootfiles will differentiate based on the script name whether
to target the encrypted or non-encrypted partittions for file installation.

Changelog-entry: customize hostOS update script for secure boot
Signed-off-by: Alex Gonzalez <alexg@balena.io>