systemd: configure panic on hung tasks and extend timeout
On long run reboot tests we have seen devices get stuck on reboot
when systemd-shutdown hangs and systemd is still refreshing the
hardware watchdog. Those devices are unreachable until power cycled.

To avoid this, we configure the devices to panic on hung task detection
which will reboot the device. However, as we know the engine might hang
itself when pulling big images, we also extend the timeout from 2m to 4m
to try and avoid this.

Changelog-entry: configure panic on hung tasks and extend timeout
Signed-off-by: Alex Gonzalez <alexg@balena.io>