meta-balena-raspberrypi:layer.conf: Fix revpi build after kirkstone update
The kirkstone upstream BSP now adds some device trees and device tree overlays
by default for all boards but those added files are not present in the older
Revpi Connect and Core 3 kernel versions we still use. To fix it, we remove
from the list these files for these 2 boards.

Signed-off-by: Florin Sarbu <florin@balena.io>