rpi-config: replace otg_mode with the dwc2 driver
The `otg_mode=1` setting makes the firmware mux the USB port
to the XHCI host-only controller. However, even though this correctly
enabled host mode in the USB port, the kernel drivers oops when a cable
is plugged into the OTG port in the ioboard, probably because the XCHI
controller does not support OTG or device mode.

This commit reverts to using the dwc2 kernel overlay which has less
performance but does not crash.

Changelog-entry: use dwc2 driver for USB host mode in CM4-SB.
Signed-off-by: Alex Gonzalez <alexg@balena.io>