linux-raspberrypi: Add back linux-raspberrypi_5.15.bbappend
The Unipi Raspberry Pi boards we support need the unipi-kernel-modules
recipe to compile against the kernel but these modules do not compile
against kernel version 6 so until we hear back from the vendor we keep
the 2 boards on the 5.15 kernel.

Also we do not apply 0006-overlays-Add-Hyperpixel4-overlays.patch and 0001-Add-tpm-slb9670-tis-spi-DT-overlay.patch
for the 5.15 anymore since the requests for them were not for the Unipi boards.

Changelog-entry: Pin the Unipi boards to kernel 5.15 because of older unipi-kernel-modules
Signed-off-by: Florin Sarbu <florin@balena.io>