rpi-cmdline: Limit earlycon configuration to Pi4 family
At this moment, when building with OS_DEVELOPMENT=1, earlycon is added
to kernel command line, however the current configuration is only
valid on the Pi4 family. Using it with any other device type will make
the kernel hang.

This patch limits the configuration to Pi4, Pi400 and CM4.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>