Add support for CM5 and RPi500
At this moment the RPi5 image does not build properly, and CM5
and RPi500 are not supported at all.

This patch does several things:
* Bumps kernel to 6.6 on the Pi5 image
* Bumps the balena bootloader kernel to 6.6
* Adds CM5 and RPi500 device trees to the RPi5 image

It was necessary to backport the 6.6 recipe from meta-raspberrypi,
as 6.6 is only available from scarthgap up, and balena-raspberrypi
is still on kirkstone.

RPi5, RPi500 and CM5 are therefore supported by a single device type.

Change-type: minor
Changelog-entry: Add support for CM5 and RPi500
Signed-off-by: Michal Toman <michalt@balena.io>