balena-image: increase boot partition size
When performing a hostOS update we perform atomic writes of individual
files, that is, first we copy a new file, and then we rename.

The rename operation is atomic on some filesystems like ext4, but on
vfat it has not been atomic until Linux kernel v6.0.

Even if probably no longer needed with >v6.0 kernels, we still use fatrw
to keep a checksummed copy of the file.

The boot.img that contains the kernel and initramfs has increased in size
and two copies of it no longer fits in the boot partition.

Changelog-entry: increase the boot partition size for CM4 to support secure boot
Signed-off-by: Alex Gonzalez <alexg@balena.io>