rpi-config: raspberrypicm4-ioboard-sb: remap audio pins
Previously the audio driver was directly called from the device tree, and
the audio pins were owned by the audio driver.

Hence, unloading the audio driver would free pins 40/41 and they could be
dynamically assigned to SPI0 and re-used to program the EEPROM.

On recent kernels the audio driver is now a platform driver called from
the vchiq driver (message queuing to the video core), and the audio pins
are now owned by the vchiq driver which cannot be unloaded. Hence, these
pins cannot be dynamically re-assigned.

This commits remaps the audio pins to GPIOs 12/13 freeing up GPIOs 40/41
to use for re-programming. This is done only for the
raspberrypicm4-ioboard-sb given that it's a new board and we do not have
to worry about existing applications.

For other device types, in particular the Raspberrypi 4 that uses by
default pins 40/41 for the headphones output, this won't work. We will
need to move to self-update support, although that still requires to
re-program the EEPROM to configure it.

Changelog-entry: remap audio pins for raspberrypicm4-ioboard-sb
Signed-off-by: Alex Gonzalez <alexg@balena.io>