Update to Scarthgap
With the update to Scarthgap we need to bring in kernel recipe 5.15 from the kirkstone BSP
layer since this version is not available anymore in the scarthgap BSP release. This should
be temporary until we update the boards that still use kernel 5.15 to a 6.x kernel version.

Changelog-entry: Update to Scarthgap
Signed-off-by: Florin Sarbu <florin@balena.io>