raspberrypicm4-ioboard-sb: use raspberrypicm4-ioboard machine
The deployment slug will keep being raspberrypicm4-ioboard-sb until we
can integrate the secure boot changes in a single image, but the machine
reverts to raspberrypicm4-ioboard as the hardware is the same and the
secure boot is just a software feature.

Aprt from the coffee file, this updates to workflows to specify a
templates path.

Changelog-entry: split secure boot and disk encryption support to a layer
Signed-off-by: Alex Gonzalez <alexg@balena.io>