layer.conf: Use full version of cyfmac43455-sdio firmware
At this moment we're defaulting to the minimal variant
of the cyfmac43455-sdio firmware on all devices except Pi5. We inherit
this from upstream, who switched to minimal in Jan 2025.

This patch makes all the devices ship default to the full variant,
which makes it possible to connect to WPA3 networks.

Changelog-entry: Default to wifi firmware with WPA3 support
Signed-off-by: Michal Toman <michalt@balena.io>