raspberrypi0-2w-64: Use kernel 6.12
When updating the kernel to 6.12, wifi on the Zero 2W wasn't working,
so we kept it at 5.15. This is because the new kernel loads
a different wifi firmware.

This patch adds the missing firmware and removes the kernel version
override, so that the Zero 2W can use the same 6.12 kernel as all
the other boards.

Changelog-entry: Update kernel to 6.12 on Raspberry Pi Zero 2W
Signed-off-by: Michal Toman <michalt@balena.io>